Move tests for new functions into appropriate files.
